requirements.txt
python3-pika
Flask==3.1.2
SQLAlchemy==2.0.43
pydantic==2.11.9
pydantic-settings==2.10.1
uwsgi==2.0.30
python-dotenv==1.0.1
psycopg2-binary==2.9.9
requests==2.32.3
PyYAML==6.0.2
pika==1.3.2
GDAL==3.10.3